+++ YUproject/iot_pothole/src/init/__init__.py
@@ -1,0 +1,2 @@
+# C:/iot_pothole/src/init/__init__.py
+__all__ = ['server', 'variable', 'web', 'result']
