+++ YUproject/iot_pothole/src/static/js/scripts.js
@@ -1,0 +1,7 @@
+/*!
+* Start Bootstrap - Bare v5.0.7 (https://startbootstrap.com/template/bare)
+* Copyright 2013-2021 Start Bootstrap
+* Licensed under MIT (https://github.com/StartBootstrap/startbootstrap-bare/blob/master/LICENSE)
+*/
+// This file is intentionally blank
+// Use this file to add JavaScript to your project
