+++ YUproject/iot_pothole/src/templates/base.html
@@ -1,0 +1,73 @@
+<!-- 
+    Jinja2라는 템플릿 엔진을 이용하여 애플리케이션 내 변수와 반복문, 조건문 등을 포함하는 
+    html 파일을 렌더링 할 수 있음.
+    < html 디자인 구상 >
+    - 일단 표를 이용하여 위치 주소와 상세 위치 버튼이 뜨게 만듬.
+    - 상세 위치 버튼을 눌렀을 때, 포트홀 이미지와 지도 상 위치가 뜸.
+
+    [ 템플릿 상속 ]
+    - 웹 사이트 레이아웃의 일관성(Consistency) 유지 및 header와 footer를 여러 곳에 사용하기 위해서
+      템플릿 상속 기능 사용
+    
+-->
+<!--
+    base.html
+    부모 html 
+-->
+<!DOCTYPE html>
+<html lang="en">
+    <head>
+        <meta charset="utf-8" />
+        <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no" />
+        <meta name="description" content="" />
+        <meta name="author" content="" />
+        <title>Pothole Management Site</title>
+        <!-- Favicon-->
+        <link rel="icon" type="image/x-icon" href="../../static/assets/favicon.ico" />
+        <!-- Core theme CSS (includes Bootstrap)-->
+        <link href="../../static/css/styles.css" rel="stylesheet" />
+    </head>
+    <body>
+        <!-- Respon-->
+        <nav class="navbar navbar-expand-lg navbar-dark bg-dark">
+            <div class="container">
+                <a class="navbar-brand" href="#">Pothole Management Site</a>
+                <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation"><span class="navbar-toggler-icon"></span></button>
+                <div class="collapse navbar-collapse" id="navbarSupportedContent">
+                    <ul class="navbar-nav ms-auto mb-2 mb-lg-0">
+                        <li class="nav-item"><a class="nav-link active" aria-current="page" href="#">Home</a></li>
+                        <li class="nav-item"><a class="nav-link" href="#">Link</a></li>
+                        <li class="nav-item dropdown">
+                            <a class="nav-link dropdown-toggle" id="navbarDropdown" href="#" role="button" data-bs-toggle="dropdown" aria-expanded="false">Dropdown</a>
+                            <ul class="dropdown-menu dropdown-menu-end" aria-labelledby="navbarDropdown">
+                                <li><a class="dropdown-item" href="#">Action</a></li>
+                                <li><a class="dropdown-item" href="#">Another action</a></li>
+                                <li><hr class="dropdown-divider" /></li>
+                                <li><a class="dropdown-item" href="#">Something else here</a></li>
+                            </ul>
+                        </li>
+                    </ul>
+                </div>
+            </div>
+        </nav>
+        <!-- Page content-->
+        <div class="container">
+            <div>
+                <h1> POTHOLE DATA LIST <h1>
+            </div>
+            {% block content %}
+            <!--
+            <div class="text-center mt-5">
+                <a href="/"><h1>검색화면 바로가기</h1></a>
+                <p class="lead">A complete project boilerplate built with Bootstrap</p>
+                <p>Bootstrap v5.1.3</p>
+            </div>
+            -->
+            {% endblock %}
+        </div>
+        <!-- Bootstrap core JS-->
+        <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/js/bootstrap.bundle.min.js"></script>
+        <!-- Core theme JS-->
+        <script src="../../static/js/scripts.js"></script>
+    </body>
+</html>
