+++ YUproject/iot_pothole/src/templates/pothole_server.html
@@ -55,7 +55,7 @@
              ">
              <p
              style="text-align:center;">
-             Pothole location imformation
+             Pothole Management Service
             </p>
              </span>
         </header>
@@ -73,7 +73,7 @@
 
 
         <footer>
-          <img src="{{url_for('static', filename='KakaoTalk_20211229_182017805.jpg')}}"
+          <img src="/static/KakaoTalk_20211229_182017805.jpg"
            width="100%"
            height="500">
           </footer>
