+++ YUproject/iot_pothole/src/templates/result/details/test.html
@@ -1,0 +1,76 @@
+<!--
+    #
+    # filename : pothole_server.html
+    # 검색한 지역의 포트홀 데이터를 띄우는 웹
+    # 너무 가라가 심해서 공부 후 업데이트 예정
+    # author : 이수영 ( 2021.12.28 )
+    # - version 1.0 (2021.12.28)
+    #   div의 id가 바뀌지 않으면 여러 맵을 띄울 수 없음.
+    #   그래서 map을 임의로 여러개 만들고 var를 이용하여 증가하면서 넣는 방법을 사용
+    #   공부 후 업데이트 필요
+    #   
+--> 
+<!DOCTYPE html> <html>
+    <head> 
+        <title>
+        GNSS Application with Kakao Map 
+        </title>
+        <meta charset = "utf-8"/> 
+        <script type="text/javascript" src="//dapi.kakao.com/v2/maps/sdk.js?appkey=3fb02947329abea5bae860500e4bdfc0"></script>
+        <style>
+        .overlaybox {position:relative;width:1000px;height:400px; padding:15px 10px;}
+        .overlaybox .boxtitle {color:#000;font-size:16px;background-size: 100px; background-color: #fff; font-weight:bold; margin-bottom:8px;}
+        table {
+        width: 100%;
+        border: 1px solid #444444;
+        }
+        th, td {
+        border: 1px solid #444444;
+        }
+        </style>
+    </head>
+<body></body>
+<div>
+    <h1> POTHOLE DATA RESULT <h1>
+</div>
+<div>
+    <script>var count=0;</script>
+    {% if data %}
+        <div id="map" style="width:70%;height:800px;"></div> <!-- 주소를 이용해서 kakao map을 띄우려면 먼저 선언된 map id가 필요함. -->
+        <script> // 자바 스크립트에서는 동적으로 만들 수 있어서 map id를 바꿔가면서 출력.
+            var addr = "{{data[3]}}"; // address
+            var content = '<div class="overlaybox">'+ // overlaybox를 표시할때 안의 내용
+            '   <div class="boxtitle">'+addr+'</div>'+
+            '   <div><img src="{{ url_for("static", filename=data[0])}}"></div>'+
+            '</div>';
+            var id_name = "map"; // id를 계속 바꾸기 위함
+            var lat = "{{data[1]|float}}";
+            var lng = "{{data[2]|float}}";
+            var mapContainer = document.getElementById(id_name), // 지도를 표시할 div 
+                mapOption = { 
+                    center: new kakao.maps.LatLng(lat, lng), // 지도의 중심좌표
+                    level: 2 // 지도의 확대 레벨
+                };
+            var position = new kakao.maps.LatLng(lat, lng); // overlay 좌표
+            var markerPosition = new kakao.maps.LatLng(lat, lng); // marker 좌표, 파란색으로 찍어주는 것
+            var marker = new kakao.maps.Marker({ // marker 객체 생성, 객체 생성 시 좌표와 함꼐 생성해서 그 위치에 생기게 하는 것.
+                position: markerPosition
+            });
+
+            var map = new kakao.maps.Map(mapContainer, mapOption);// map 객체 생성, 맵 객체 생성할 때, 지도가 들어가는 컨테이너랑 지도의 중심좌표, 확대 레벨을 넣어서 생성.
+            var customOverlay = new kakao.maps.CustomOverlay({// overlay 객체 생성, 포트홀 이미지와 주소를 넣을 수 있게 하는 객체
+                position: position,
+                content: content,
+                xAnchor: 0.3,
+                yAnchor: 0.91
+            });
+            customOverlay.setMap(map);     
+            marker.setMap(map);
+        </script>
+    {% else %}
+        <p>해당 지역에는 포트홀이 존재하지 않습니다.</p>
+    {% endif %}
+
+</div>
+</body>
+</html>
