+++ YUproject/iot_pothole/src/templates/result/new.html
@@ -1,0 +1,17 @@
+{% extends "base.html" %}
+
+{% block content %}
+    <div>
+        <a href="/"><h3>검색화면 바로가기</h3></a>
+    </div>
+    {% if data %}
+        <ul>
+        {% for pothole_data in data %}
+            <li><a href="/result/details/{{ pothole_data[0] }}/{{ pothole_data[1] }}/{{ pothole_data[2] }}/{{ pothole_data[3] }}/">{{ pothole_data[3] }}</a></li>
+        {% endfor %}
+        </ul>
+    {% else %}
+        <p>해당 지역에는 포트홀이 존재하지 않습니다.</p>
+    {% endif %}
+{% endblock %}
+<!--details/{{ pothole_data.img_path }}/{{ pothole_data.latitude }}/{{ pothole_data.longitude }}/{{ pothole_data.addr }}/-->
